Make Constraint and Conshdlr generic over data
It defaults to None(Type), which means all uses as parameter should
include the explicit TypeVar to accept everything. The None default is
because when creating a constraint, its data is None unless it gets set
afterwards.

By the way, I also discovered that the data on the Python constraint is
not synced with the data on the C constraint, e.g. for disjunction
constraints the data needs a specific format to format, but from the
Python side it looks like any other Constraint with no associated data.